Adjusted to latest shell
