+++ __mocks__/openmrs-esm-framework.mock.tsx
@@ -1,5 +1,22 @@
 import React from 'react';
+import { of, never } from 'rxjs';
 import { ConfigMock } from './chart-widgets-config.mock';
+
+export function openmrsFetch() {
+  return new Promise(() => {});
+}
+
+export function openmrsObservableFetch() {
+  return of({ data: { entry: [] } });
+}
+
+export function UserHasAccessReact(props: any) {
+  return props.children;
+}
+
+export function createErrorHandler() {
+  return jest.fn().mockReturnValue(never());
+}
 
 export function defineConfigSchema() {}
 
@@ -17,3 +34,5 @@ export function useConfig() {
 export const ComponentContext = React.createContext({
   moduleName: 'fake-module-config',
 });
+
+export const showToast = jest.fn();
