+++ package.json
@@ -16,7 +16,7 @@
     "prettier": "prettier --config prettier.config.js --write \"src/**/*.{ts,tsx}\"",
     "typescript": "tsc",
     "prepublishOnly": "npm run build",
-    "test": "jest",
+    "test": "jest --passWithNoTests",
     "coverage": "npm test -- --coverage"
   },
   "browserslist": [
