+++ src/medications-summary/floating-order-basket-button.component.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
-import { Button, Tag } from 'carbon-components-react';
+import Button from 'carbon-components-react/es/components/Button';
+import Tag from 'carbon-components-react/es/components/Tag';
+import styles from './floating-order-basket-button.scss';
 import { ShoppingBag16 } from '@carbon/icons-react';
 import { useTranslation } from 'react-i18next';
-import styles from './floating-order-basket-button.scss';
 import { connect } from 'unistore/react';
 import { switchTo } from '@openmrs/esm-framework';
 import { OrderBasketStoreActions, OrderBasketStore } from '../order-basket-store';
