+++ src/medications-summary/medications-summary.component.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { useTranslation } from 'react-i18next';
+import DataTableSkeleton from 'carbon-components-react/es/components/DataTableSkeleton';
 import FloatingOrderBasketButton from './floating-order-basket-button.component';
 import styles from './medications-summary.scss';
 import MedicationsDetailsTable from '../components/medications-details-table.component';
-import { DataTableSkeleton } from 'carbon-components-react';
+import { useTranslation } from 'react-i18next';
 import { usePatientOrders } from '../utils/use-current-patient-orders.hook';
 
 export interface MedicationsSummaryProps {
