+++ src/order-basket/order-basket-search.component.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
-import { Search } from 'carbon-components-react';
-import { useTranslation } from 'react-i18next';
+import Search from 'carbon-components-react/es/components/Search';
 import styles from './order-basket-search.scss';
 import OrderBasketSearchResults from './order-basket-search-results';
+import { useTranslation } from 'react-i18next';
 import { OrderBasketItem } from '../types/order-basket-item';
 
 export interface OrderBasketSearchProps {
@@ -19,7 +19,7 @@ export default function OrderBasketSearch({ encounterUuid, onSearchResultClicked
       <div className={styles.searchPopupContainer}>
         <Search
           value={searchTerm}
-          placeHolderText={t('searchFieldPlaceholder', 'Search for an order (e.g. "Aspirin")')}
+          placeholder={t('searchFieldPlaceholder', 'Search for an order (e.g. "Aspirin")')}
           labelText={t('searchFieldPlaceholder', 'Search for an order (e.g. "Aspirin")')}
           onChange={e => setSearchTerm(e.currentTarget?.value ?? '')}
         />
